+++ CHANGELOG.md
@@ -24,6 +24,15 @@
   overlay create` command. The overlay will be applied read-only, by default,
   when executing the OCI-SIF. To write changes to the container into the overlay,
   use the `--writable` flag.
+- A writable overlay is added to an OCI-SIF file as an ext3 format layer,
+  appended to the encapsulated OCI image. After the overlay has been modified,
+  use the `singularity overlay sync` command to synchronize the OCI digests with
+  the overlay content.
+- A new `singularity overlay seal` command converts a writable overlay inside
+  an OCI-SIF image into a read-only squashfs layer. This seals changes made to
+  the image via the overlay, so that they are permanent.
+- Added a new `instance run` command that will execute the runscript when an
+  instance is initiated instead of executing the startscript.
 
 ### Bug Fixes